Refractor code
