+++ src/main/java/com/dawn/jat/illuminati/post/advice/PostExceptionAdvice.java
@@ -1,10 +1,8 @@
 package com.dawn.jat.illuminati.post.advice;
 
+import com.dawn.jat.illuminati.core.response.ErrorResponse;
 import com.dawn.jat.illuminati.post.exception.PostNotFoundException;
 import com.dawn.jat.illuminati.post.exception.PostSummaryNotFoundException;
-import com.dawn.jat.illuminati.post.response.ErrorResponse;
-import java.util.LinkedHashMap;
-import java.util.Map;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
@@ -25,12 +23,7 @@ public class PostExceptionAdvice {
      */
     @ExceptionHandler(PostSummaryNotFoundException.class)
     public ResponseEntity<ErrorResponse> exceptionPostSummary(Exception exception) {
-        Map<String, Object> err = new LinkedHashMap<>();
-        ErrorResponse error = new ErrorResponse();
-        err.put("code", HttpStatus.NO_CONTENT.value());
-        err.put("message", exception.getMessage());
-        error.setError(err);
-        error.setSuccess(false);
+        ErrorResponse error = new ErrorResponse(HttpStatus.NO_CONTENT, exception);
         return new ResponseEntity<>(error, HttpStatus.OK);
     }
 }
