--- src/main/java/com/dawn/jat/illuminati/post/response/ErrorResponse.java
@@ -1,13 +1,0 @@
-package com.dawn.jat.illuminati.post.response;
-
-import java.util.Map;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class ErrorResponse {
-    private Map<String, Object> error;
-    private Boolean success;
-}
