--- src/main/java/com/dawn/jat/illuminati/post/response/SuccessResponse.java
@@ -1,13 +1,0 @@
-package com.dawn.jat.illuminati.post.response;
-
-import java.util.List;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class SuccessResponse {
-    private List<?> data;
-    private Boolean success;
-}
