+++ src/test/java/com/dawn/jat/illuminati/core/response/SuccessResponseTest.java
@@ -1,13 +1,13 @@
-package com.dawn.jat.illuminati.post.response;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
+package com.dawn.jat.illuminati.core.response;
 
 import com.dawn.jat.illuminati.post.entity.PostSummaryEntity;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+
 import java.util.Arrays;
 import java.util.List;
 
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class SuccessResponseTest {
     private static PostSummaryEntity postSummaryEntity;
@@ -26,10 +26,9 @@ public class SuccessResponseTest {
     void success_Response_returnGetPostSuccessStatus() {
         List<PostSummaryEntity> mockPostSumEntities = Arrays.asList(postSummaryEntity);
 
-        SuccessResponse succResp = new SuccessResponse();
-        succResp.setSuccess(true);
+        SuccessResponse succResp = new SuccessResponse(mockPostSumEntities);
+
         assertEquals(true, succResp.getSuccess());
-        succResp.setData(mockPostSumEntities);
         assertEquals(mockPostSumEntities, succResp.getData());
     }
 }
